+++ lib/language_pack/rails3.rb
@@ -6,9 +6,10 @@ class LanguagePack::Rails3 < LanguagePack::Rails2
   # detects if this is a Rails 3.x app
   # @return [Boolean] true if it's a Rails 3.x app
   def self.use?
-    super &&
-      File.exists?("config/application.rb") &&
-      File.read("config/application.rb").include?("Rails::Application")
+    if gemfile_lock?
+      rails_version = LanguagePack::Ruby.gem_version('railties')
+      rails_version >= Gem::Version.new('3.0.0') && rails_version < Gem::Version.new('4.0.0') if rails_version
+    end
   end
 
   def name